backend/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.1
